requirements.txt
fastapi==0.101.1
kaleido==0.2.1
pandas==2.0.3
pytz==2023.3
requests==2.31.0
uvicorn==0.23.2
plotly==5.16.1
PyYAML==6.0.1
